+++ Assets/01.Scripts/Ingame/Click/ClickInfo.cs
@@ -3,12 +3,12 @@ using UnityEngine;
 public struct ClickInfo
 {
     public Vector2 WorldPosition;
-    public int Damage;
+    public double Damage;
     public int ToolLevel;
     public bool IsAutoClick;
-    public int Reward;  // 획득 재화
+    public double Reward;  // 획득 재화
 
-    public ClickInfo(Vector2 worldPosition, int damage, int toolLevel, bool isAutoClick)
+    public ClickInfo(Vector2 worldPosition, double damage, int toolLevel, bool isAutoClick)
     {
         WorldPosition = worldPosition;
         Damage = damage;
